Add TCP_MPTCPIFY socket option for MPTCP support
Multipath TCP (MPTCP), as standardized in RFC 8684, is a major
evolution of the TCP protocol. It enables a transport connection to
use multiple network paths simultaneously for redundancy, resilience,
and bandwidth aggregation. Since its introduction in Linux kernel v5.6,
it has become a key technology for modern networking, particularly on
mobile devices with multiple interfaces (e.g., Wi-Fi and cellular).
For more details, see the project website: https://mptcp.dev.

This commit introduces the TCP_MPTCPIFY (the name was chosen to align
with 'mptcpize' tool in mptcpd and 'mptcpify' in BCC) socket option to
enable MPTCP support on Linux systems. The implementation includes:

- A new native method mptcpify0 in LinuxSocketOptions.c that creates
  an MPTCP socket and duplicates it to the original file descriptor,
  effectively converting the socket to use MPTCP. Relevant socket
  options are copied from the original socket to the new MPTCP socket.
- The TCP_MPTCPIFY option is exposed via ExtendedSocketOptions and is
  available only on Linux platforms.
- The feature is conditionally compiled using IPPROTO_MPTCP where
  supported.

This option must be set before calling bind() or connect() to take
effect.

Suggested-by: Alan Bateman <alan.bateman@oracle.com>
Co-developed-by: Gang Yan <yangang@kylinos.cn>
Signed-off-by: Gang Yan <yangang@kylinos.cn>
Co-developed-by: Xiang Gao <gaoxiang@kylinos.cn>
Signed-off-by: Xiang Gao <gaoxiang@kylinos.cn>
Signed-off-by: Geliang Tang <geliang@kernel.org>