Add test cases for MPTCP socket functionality
These tests verify the bahavior of the new MPTCP-enabled socket
constructors:
- Socket(host, port, stream, mptcp)
- ServerSocket(port, backlog, address, mptcp)

When the Linux kernel enables MPTCP, the sockets can be created
successfully and data can be transfered over a loopback connection.
When the MPTCP is disabled, the sockets API fails with the IOException
which is expected.

The test design follows OpenJDK jtreg conventions:
- @requires os.family linux.
- add MptcpUtil to detect kernel support for common use.
- Run server and client in one jvm, with out user interaction.

Co-developed-by: Gang Yan <yangang@kylinos.cn>
Signed-off-by: Gang Yan <yangang@kylinos.cn>
Co-developed-by: Geliang Tang <geliang@kernel.org>
Signed-off-by: Geliang Tang <geliang@kernel.org>
Signed-off-by: Xiang Gao <gaoxiang@kylinos.cn>